backend/requirements.txt
# Django backend requirements
Django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework.authtoken>=3.14
django-cors-headers>=4.3
Pillow>=10.0
psycopg2-binary>=2.9
python-dotenv>=1.0.0
gunicorn>=21.0.0
